src/llm-api/pyproject.toml
[project]
name = "consult-llm-api"
version = "0.1.0"
description = "FastAPI proxy that calls Vertex AI Gemini for The Consult."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.0",
    "google-genai>=1.39.1",
    "python-dotenv>=1.0.1",
    "chromadb>=0.5.20",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.2",
    "pytest>=8.3.4",
    "black>=24.1.1",
    "flake8>=7.0.0",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["api*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
